Experimental fix for heap error on infinite write
For some reason, simply inserting a statement that prints out something
to the serial console prevents a heap error. However, running it for
long periods of time may result in a heap error, as this may only
delay it. I think that the compositor server isn't ready when it
receives the next packet, which can cause the invalid magic byte errors
seen before it crashes. This may help #4