Fix and build flock
Flock system call is fixed in this. The cause was an incompatible
declaration, with the flock call in newlib with flock(int, int),
but our implementation flock(int, unsigned int). This changes the
declarations to the correct ones, which fixes the build. Flock
is reenabled in this, so flock-test works again.